requirements.txt
Flask
SQLAlchemy
opencv-python
numpy
pillow
python-dotenv
cryptography
gunicorn
pytest
liboqs-python
